requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
pandas==2.2.3
numpy==1.26.4
google-generativeai==0.8.3
pinecone==3.0.2
beautifulsoup4==4.12.3
rank-bm25==0.2.2
httpx==0.27.2
tqdm==4.66.5
